Refactor: scrape function => scrape folder
